feat: root.vue shared layout
