+++ apps/frontend/src/pages/hosting/manage/[id]/index.vue
@@ -1,19 +1,6 @@
 <script setup lang="ts">
 import { injectModrinthServerContext, ServersManageOverviewPage } from '@modrinth/ui'
-import type { ServerState, Stats } from '@modrinth/utils'
 
-type ServerProps = {
-	isConnected: boolean
-	isWsAuthIncorrect: boolean
-	stats: Stats
-	serverPowerState: ServerState
-	powerStateDetails?: {
-		oom_killed?: boolean
-		exit_code?: number
-	}
-}
-
-const props = defineProps<ServerProps>()
 const { server } = injectModrinthServerContext()
 
 useHead({
@@ -22,5 +9,5 @@ useHead({
 </script>
 
 <template>
-	<ServersManageOverviewPage v-bind="props" />
+	<ServersManageOverviewPage />
 </template>
