+++ packages/ui/src/components/servers/index.ts
@@ -8,5 +8,6 @@ export { default as SaveBanner } from './SaveBanner.vue'
 export * from './server-header'
 export { default as ServerListEmpty } from './server-list-empty/ServerListEmpty.vue'
 export { type PendingChange, default as ServerListing } from './ServerListing.vue'
+export { default as ServerSettingsModal } from './ServerSettingsModal.vue'
 export { default as ServerSetupModal } from './ServerSetupModal.vue'
 export { default as ServersPromo } from './ServersPromo.vue'
