+++ packages/ui/src/components/servers/marketing/MedalServerCountdown.vue
@@ -34,18 +34,27 @@
 			<button class="z-10 my-auto" @click="openUpgradeModal"><RocketIcon /> Upgrade</button>
 		</ButtonStyled>
 	</div>
-	<ServersUpgradeModalWrapper ref="upgradeModal" />
+	<ServersUpgradeModalWrapper
+		ref="upgradeModal"
+		:stripe-publishable-key="props.stripePublishableKey ?? ''"
+		:site-url="props.siteUrl ?? ''"
+		:products="props.products ?? []"
+	/>
 </template>
 
 <script setup lang="ts">
+import type { Labrinth } from '@modrinth/api-client'
 import { ClockIcon, RocketIcon } from '@modrinth/assets'
-import { ButtonStyled, MedalBackgroundImage } from '@modrinth/ui'
-import type { UserSubscription } from '@modrinth/utils'
+import { useQuery } from '@tanstack/vue-query'
 import dayjs from 'dayjs'
 import dayjsDuration from 'dayjs/plugin/duration'
-import type { ComponentPublicInstance } from 'vue'
+import { type ComponentPublicInstance, computed, onMounted, onUnmounted, ref } from 'vue'
 
-import ServersUpgradeModalWrapper from '../ServersUpgradeModalWrapper.vue'
+import ButtonStyled from '#ui/components/base/ButtonStyled.vue'
+import ServersUpgradeModalWrapper from '#ui/components/billing/ServersUpgradeModalWrapper.vue'
+import { injectModrinthClient } from '#ui/providers'
+
+import MedalBackgroundImage from './MedalBackgroundImage.vue'
 
 dayjs.extend(dayjsDuration)
 
@@ -54,15 +63,16 @@ const upgradeModal = ref<UpgradeWrapperRef | null>(null)
 
 const props = defineProps<{
 	serverId?: string
+	stripePublishableKey?: string
+	siteUrl?: string
+	products?: Labrinth.Billing.Internal.Product[]
 }>()
 
-const { data: subscriptions } = await useLazyAsyncData(
-	'countdown-subscriptions',
-	() =>
-		useBaseFetch(`billing/subscriptions`, {
-			internal: true,
-		}) as Promise<UserSubscription[]>,
-)
+const client = injectModrinthClient()
+const { data: subscriptions } = useQuery({
+	queryKey: ['billing', 'subscriptions'],
+	queryFn: () => client.labrinth.billing_internal.getSubscriptions(),
+})
 
 const expiryDate = computed(() => {
 	for (const subscription of subscriptions.value || []) {
