+++ packages/ui/src/components/servers/marketing/index.ts
@@ -1,2 +1,3 @@
 export { default as MedalBackgroundImage } from './MedalBackgroundImage.vue'
+export { default as MedalServerCountdown } from './MedalServerCountdown.vue'
 export { default as MedalServerListing } from './MedalServerListing.vue'
