+++ packages/ui/src/composables/index.ts
@@ -13,4 +13,6 @@ export * from './server-console'
 export * from './server-manage-core-runtime'
 export * from './sticky-observer'
 export * from './terminal'
+export * from './use-server-image'
+export * from './use-server-project'
 export * from './virtual-scroll'
