+++ packages/ui/src/composables/server-manage-core-runtime.ts
@@ -346,8 +346,12 @@ export function useServerManageCoreRuntime(options: UseServerManageCoreRuntimeOp
 		worldId: options.worldId as Ref<string | null>,
 		server: options.server as Ref<Archon.Servers.v0.Server>,
 		isConnected,
+		isWsAuthIncorrect,
 		powerState: serverPowerState,
+		powerStateDetails,
 		isServerRunning,
+		stats,
+		uptimeSeconds,
 		backupsState,
 		markBackupCancelled,
 		isSyncingContent: options.isSyncingContent as Ref<boolean>,
