+++ packages/ui/src/composables/use-server-project.ts
@@ -1,17 +1,18 @@
 import type { Archon } from '@modrinth/api-client'
-import type { Project } from '@modrinth/utils'
 import { useQuery } from '@tanstack/vue-query'
-import { $fetch } from 'ofetch'
 import { computed, type ComputedRef } from 'vue'
+
+import { injectModrinthClient } from '#ui/providers'
 
 // TODO: Remove and use v1
 export function useServerProject(
 	upstream: ComputedRef<Archon.Servers.v0.Server['upstream'] | null>,
 ) {
+	const client = injectModrinthClient()
+
 	return useQuery({
 		queryKey: computed(() => ['servers', 'project', upstream.value?.project_id ?? null]),
-		queryFn: () =>
-			$fetch<Project>(`https://api.modrinth.com/v2/project/${upstream.value!.project_id}`),
+		queryFn: () => client.labrinth.projects_v2.get(upstream.value!.project_id!),
 		enabled: computed(() => !!upstream.value?.project_id),
 	})
 }
