+++ packages/ui/src/layouts/wrapped/index.ts
@@ -4,3 +4,4 @@ export { default as ServersManageContentPage } from './hosting/manage/content.vu
 export { default as ServersManageFilesPage } from './hosting/manage/files.vue'
 export { default as ServersManagePageIndex } from './hosting/manage/index.vue'
 export { default as ServersManageOverviewPage } from './hosting/manage/overview.vue'
+export { default as ServersManageRootLayout } from './hosting/manage/root.vue'
